(tae): small code refactor
