+++ pkg/vm/engine/tae/dataio/segmentio/segment_test.go
@@ -18,7 +18,7 @@ import (
 	"path"
 	"testing"
 
-	"github.com/matrixorigin/matrixone/pkg/vm/engine/aoe/storage/common"
+	"github.com/matrixorigin/matrixone/pkg/vm/engine/tae/common"
 	"github.com/matrixorigin/matrixone/pkg/vm/engine/tae/testutils"
 	"github.com/stretchr/testify/assert"
 )
